+++ src/components/cc.js
@@ -1,0 +1,1 @@
+// http://api.fixer.io/latest?base=USD&symbols=EUR,RUB,GBP